requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2

# Computer Vision
opencv-python==4.8.1.78
pillow==10.1.0
scikit-image==0.22.0

# Deep Learning - Vision Transformers
torch==2.1.1
torchvision==0.16.1
timm==0.9.12  # Transformers for PyTorch Image Models (ViT support)

# Data processing
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4
scikit-learn==1.3.2

# Image augmentation
albumentations==1.3.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Configuration and utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
pyyaml==6.0.1

# Logging and monitoring
tqdm==4.66.1
colorlog==6.8.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools (optional)
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Additional utilities
click==8.1.7
requests==2.31.0
psutil==5.9.6